Dups protection

+++ inc/Server.hpp
@@ -16,7 +16,7 @@ class Server
 	private:
 		int							_listen;
 		std::vector<std::string>	_server_name;
-		int							_client_max_body_size;
+		long						_client_max_body_size;
 		std::string					_root;
 		std::string					_upload_store;
 		std::map<int, std::string>	_error_page;
@@ -29,7 +29,7 @@ class Server
 		//GET & SET
 		int getListen() const;
 		std::vector<std::string> getServerName() const;
-		int getClientMaxBodySize() const;
+		long getClientMaxBodySize() const;
 		std::string getRoot() const;
 		std::string getUploadStore() const;
 		Location &getLocations(std::string const dir);
@@ -37,7 +37,7 @@ class Server
 		std::map<int, std::string> getErrorPage() const;
 		void setListen(int listen_);
 		void setServerName(const std::vector<std::string> &serverName);
-		void setClientMaxBodySize(int clientMaxBodySize);
+		void setClientMaxBodySize(long clientMaxBodySize);
 		void setRoot(const std::string &root);
 		void setUploadStore(const std::string &uploadStore);
 		void setErrorPage(const std::map<int, std::string> &errorPage);
